+++ test/encode-decode.js
@@ -141,13 +141,36 @@ test('decode encode line', function (t) {
       },
       geometry: {
         type: 'LineString',
-        coordinates: [[0, 1], [1, 2], [2, 4]],
+        coordinates: [[0, 1], [1, 2], [3, 4]],
       }
     }]
   }
   const decoded = toGeoJSON(toGeorender(fc, { includeAllTags: true }))
-  // console.log(decoded.features[0].geometry)
-  t.deepEqual(decoded, fc)
+  // console.log(decoded.features[0])
+  t.deepEqual(decoded, fc, 'feature collection : line matches')
+  t.end()
+})
+
+test('decode encode multi-line', function (t) {
+  const fc = {
+    type: 'FeatureCollection',
+    features: [{
+      type: 'Feature',
+      properties: {
+        waterway: 'river',
+        meanWaterLevel: 12,
+      },
+      geometry: {
+        type: 'MultiLineString',
+        coordinates: [
+          [[0, 1], [1, 2]],
+          [[3, 4], [5, 6]]
+        ],
+      }
+    }]
+  }
+  const decoded = toGeoJSON(toGeorender(fc, { includeAllTags: true }))
+  t.notDeepEqual(decoded, fc, 'fails: MultiLineString does not match, encoding does not preserve this information, and flattens into a single LineString.')
   t.end()
 })
 
@@ -191,6 +214,34 @@ test('decode polygon area with tags', function (t) {
           type: 'multipolygon',
           place: 'city',
           population: 342_259,
+        },
+        geometry: {
+          type: 'Polygon',
+          coordinates: [
+            [[10,20],[30,40],[50,0],[10,20]]
+          ],
+        },
+      }
+    ],
+  }
+  const decoded = toGeoJSON(toGeorender(fc, { includeAllTags: true }))
+  t.deepEqual(decoded, fc)
+  t.end()
+})
+
+test('decode polygon area with tags and names', function (t) {
+  const fc = {
+    type: 'FeatureCollection',
+    features: [
+      {
+        type: 'Feature',
+        properties: {
+          type: 'multipolygon',
+          place: 'country',
+          population: 342_259,
+          name: 'Puerto Rico',
+          'name:taíno': 'Borikén',
+          'alt_name:en': 'Rich Port',
         },
         geometry: {
           type: 'Polygon',